refactor: card header file organisation and descriptionRadio story (#3842)
