+++ editor.planx.uk/src/@planx/components/DateInput/Public.tsx
@@ -1,9 +1,12 @@
 import Box from "@mui/material/Box";
 import { visuallyHidden } from "@mui/utils";
-import { DateInput, paddedDate } from "@planx/components/DateInput/model";
-import { dateRangeSchema } from "@planx/components/DateInput/model";
+import {
+  DateInput,
+  dateRangeSchema,
+  paddedDate,
+} from "@planx/components/DateInput/model";
 import Card from "@planx/components/shared/Preview/Card";
-import CardHeader from "@planx/components/shared/Preview/CardHeader";
+import { CardHeader } from "@planx/components/shared/Preview/CardHeader/CardHeader";
 import { PublicProps } from "@planx/components/ui";
 import { useFormik } from "formik";
 import React from "react";
