+++ editor.planx.uk/src/@planx/components/NextSteps/Public.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import NextStepsList from "ui/public/NextStepsList";
 
 import Card from "../shared/Preview/Card";
-import CardHeader from "../shared/Preview/CardHeader";
+import { CardHeader } from "../shared/Preview/CardHeader/CardHeader";
 import { NextSteps } from "./model";
 
 export type Props = PublicProps<NextSteps>;
