+++ editor.planx.uk/src/@planx/components/NumberInput/Public.tsx
@@ -1,5 +1,5 @@
 import Card from "@planx/components/shared/Preview/Card";
-import CardHeader from "@planx/components/shared/Preview/CardHeader";
+import { CardHeader } from "@planx/components/shared/Preview/CardHeader/CardHeader";
 import { PublicProps } from "@planx/components/ui";
 import { useFormik } from "formik";
 import isNil from "lodash/isNil";
