+++ editor.planx.uk/src/@planx/components/shared/Preview/CardHeader/types.ts
@@ -1,0 +1,9 @@
+export interface ICardHeader {
+  title?: string;
+  description?: string;
+  info?: string;
+  policyRef?: string;
+  howMeasured?: string;
+  definitionImg?: string;
+  img?: string;
+}
