+++ editor.planx.uk/src/@planx/components/shared/Radio/DescriptionRadio/DescriptionRadio.stories.tsx
@@ -1,0 +1,21 @@
+import { Meta, StoryObj } from "@storybook/react";
+
+import DescriptionRadio from "./DescriptionRadio";
+
+const meta = {
+  title: "Design System/Atoms/Form Elements/Radio/DescriptionRadio",
+  component: DescriptionRadio,
+} satisfies Meta<typeof DescriptionRadio>;
+
+type Story = StoryObj<typeof meta>;
+
+export default meta;
+
+export const Basic = {
+  args: {
+    title: "Prior approval",
+    description: "This option has a description underneath.",
+    onChange: () => console.log("hi"),
+    id: "1",
+  },
+} satisfies Story;
